refactor: add auth helpers to reduce boilerplate
- Create src/lib/auth.ts with requireAuth and requireWorkspace helpers
- Update all private pages to use shared auth helpers
- Remove duplicate auth and workspace check patterns
- Simplify server action authentication
